And yet again - a few more fixes

+++ MathsLib/Matrix3x3.cpp
@@ -427,13 +427,13 @@ namespace MathsLib
 		return vec;
 	}
 
-	float* Matrix3x3::operator[](unsigned int i)
+	float* Matrix3x3::operator[](size_t i)
 	{
 		Assert(i < 3, "Matrix row index not allowed");
 		return (_m[i]);
 	}
 
-	const float* const Matrix3x3::operator[](unsigned int i) const
+	const float* const Matrix3x3::operator[](size_t i) const
 	{
 		Assert(i < 3, "Matrix row index not allowed");
 		return (_m[i]);
