+++ MathsLib/Matrix3x3.h
@@ -213,8 +213,8 @@ namespace MathsLib
 
 		bool operator==(const Matrix3x3& _mat);
 
-		float* operator[](unsigned int i);
-		const float* const operator[](unsigned int i) const;
+		float* operator[](size_t i);
+		const float* const operator[](size_t i) const;
 
 		/// Returns array of data.
 		float* getData();
