+++ MathsLib/Vector4.h
@@ -179,8 +179,8 @@ namespace MathsLib
 		bool operator!=(const Vector4& _vec);
 		friend bool operator!=(const Vector4& _vec1, const Vector4& _vec2);
 
-		float operator[](unsigned int i);
-		float operator[](unsigned int i) const;
+		float operator[](size_t i);
+		float operator[](size_t i) const;
 
 		/// Returns array of data.
 		float* getData();
